#2 added documentation, added department class that aggregates everything and represents the whole module

+++ capitalismx-game/capitalismx-hr/src/main/java/de/uni/mannheim/capitalismx/hr/domain/JobSatisfaction.java
@@ -7,11 +7,14 @@ package de.uni.mannheim.capitalismx.hr.domain;
 public class JobSatisfaction {
 
     /**
+     * See p.26
+     * Take the 3rd root of the original scale and multiply by 0.33 (?)
      *
-     * @param originalScale
-     * @return
+     * @param originalScale the original scale that was composed of different factors that influences the
+     *                      job satisfaction
+     * @return The job satisfaction score based on the original scale
      */
-    public double getJobSatisfactionScore (int originalScale) {
+    public double getJobSatisfactionScore (double originalScale) {
         return Math.pow(originalScale, 1.0/3.0) * 0.33;
     }
 }
